Added Controller & Integration testing

+++ src/main/java/com/matera/cursoferias/petstore/business/EspecieBusiness.java
@@ -6,15 +6,20 @@ import java.util.List;
 import org.springframework.stereotype.Component;
 
 import com.matera.cursoferias.petstore.entity.Especie;
+import com.matera.cursoferias.petstore.entity.Pet;
+import com.matera.cursoferias.petstore.exception.ResourceBadRequestException;
+import com.matera.cursoferias.petstore.exception.ResourceNotFoundException;
 import com.matera.cursoferias.petstore.repository.EspecieRepository;
 
 @Component
 public class EspecieBusiness implements EspecieBusinessInterface {
 
 	private EspecieRepository especieRepository;
+	private PetBusiness petBusiness;
 	
-	public EspecieBusiness(EspecieRepository especieRepository) {
+	public EspecieBusiness(EspecieRepository especieRepository, PetBusiness petBusiness) {
 		this.especieRepository = especieRepository;
+		this.petBusiness = petBusiness;
 	}
 
 	@Override
@@ -33,11 +38,25 @@ public class EspecieBusiness implements EspecieBusinessInterface {
 
 	@Override
 	public Especie findById(Long id) {
-		return especieRepository.findById(id).orElse(null);
+		Especie especie = especieRepository.findById(id).orElse(null);
+		
+		if (especie == null) {
+			throw new ResourceNotFoundException(String.format("Espécie %d não encontrada!", id));
+		}
+		
+		return especie; 	
 	}
 
 	@Override
 	public void deleteById(Long id) {
+		findById(id);
+		
+		List<Pet> pets = petBusiness.findByEspecie_Id(id);
+		
+		if (!pets.isEmpty()) {
+			throw new ResourceBadRequestException(String.format("Espécie %d não pode ser excluída pois possui Pets!", id));
+		}
+		
 		especieRepository.deleteById(id);
 	}
